+++ src/main/scala/org/fs/checker/provider/RawProvider.scala
@@ -13,5 +13,8 @@ trait RawProvider extends GenProvider {
   def requiresAuth: Boolean
 
   /** Create a provider, initialized using a given (full) config and content dumping service */
+  @throws[IllegalArgumentException]("if tracker instance creation fails, e.g. invalid credentials" +
+    " (page content should already be dumped to disk)")
+  @throws[IllegalStateException]("if tracker is accessible but isn't working, e.g. server-side maintenance")
   def withConfig(config: Config, dumpService: PageContentDumpService): ConfiguredProvider
 }