+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### SNAPSHOT
+### 1.4.2
 * Added providers for the following trackers:
   * RuTracker (rutracker.org)
   * NoNaMe Club (nnmclub.to)